+++ libft/ft_memset.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-void    *ft_memset(void *s, int c, unsigned int n)
+void    *ft_memset(void *s, int c, size_t n)
 {
-    unsigned int i;
+    size_t i;
     unsigned char *p;
 
     i = 0;
